+++ lib/src/sdm/src/ingredient.dart
@@ -1,0 +1,41 @@
+part of '../sdm.dart';
+
+@freezed
+class SdmIngredient with _$SdmIngredient {
+  const factory SdmIngredient({
+    /// [id] of the pen entity. This ID is unique.
+    required String id,
+
+    /// [name] of the pen.
+    required String name,
+
+    /// [code] of the pen.
+    required String code,
+
+    /// [pricePerKg] is the price of the pen per kilogram.
+    double? pricePerKg,
+
+    /// [isArchived] is the status of the pen.
+    bool? isArchived,
+
+    /// [sourceId] is the device id of the pen.
+    String? sourceId,
+
+    /// [source] is the device of the pen.
+    Device? source,
+
+    /// [createdAt] is the creation date of the pen.
+    @TimestampOrNullConverter() DateTime? createdAt,
+
+    /// [createdBy] is the user id who created the pen.
+    User? createdBy,
+
+    /// [updatedAt] is the last update date of the pen.
+    @TimestampOrNullConverter() DateTime? updatedAt,
+
+    /// [updatedBy] is the user id who updated the pen.
+    User? updatedBy,
+  }) = _SdmIngredient;
+
+  factory SdmIngredient.fromJson(Map<String, dynamic> json) => _$SdmIngredientFromJson(json);
+}